requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
brotlipy=0.7.0=py39h46acfd9_1001
ca-certificates=2021.5.30=h4653dfc_0
certifi=2021.5.30=py39h2804cbe_0
cffi=1.14.5=py39h702c04f_0
chardet=4.0.0=py39h2804cbe_1
click=8.0.1=py39h2804cbe_0
cryptography=3.4.7=py39h73257c9_0
cycler=0.10.0=py_2
dataclasses=0.8=pyhc8e2a94_1
filelock=3.0.12=pyh9f0ad1d_0
freetype=2.10.4=h17b34a0_1
future=0.18.2=py39h2804cbe_3
huggingface_hub=0.0.9=pyhd8ed1ab_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=4.3.1=py39h2804cbe_0
importlib_metadata=4.3.1=hd8ed1ab_0
jbig=2.1=h3422bc3_2003
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h27ca646_0
kiwisolver=1.3.1=py39h5a63225_1
lcms2=2.12=had6a04f_0
lerc=2.2.1=h9f76cd9_0
libblas=3.9.0=9_openblas
libcblas=3.9.0=9_openblas
libcxx=11.1.0=h168391b_0
libdeflate=1.7=h27ca646_5
libffi=3.3=h9f76cd9_2
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_22
libgfortran5=11.0.1.dev0=hf114ba7_22
liblapack=3.9.0=9_openblas
libopenblas=0.3.15=openmp_hf330de4_1
libpng=1.6.37=hf7e6567_2
libprotobuf=3.15.8=hccf11d3_0
libtiff=4.3.0=hc6122e1_1
libwebp-base=1.2.0=h27ca646_2
llvm-openmp=11.1.0=hf3c4609_1
lz4-c=1.9.3=h9f76cd9_0
matplotlib=3.4.2=py39hdf13c20_0
matplotlib-base=3.4.2=py39h5aa4fe7_0
ncurses=6.2=h9aa5885_4
ninja=1.10.2=h4d860bb_0
numpy=1.20.3=py39h1f3b974_1
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=h062765e_1
openssl=1.1.1k=h27ca646_0
packaging=20.9=pyh44b312d_0
pillow=8.2.0=py39h282ff46_1
pip=21.1.2=pyhd8ed1ab_0
pycparser=2.20=pyh9f0ad1d_2
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py39h2804cbe_3
python=3.9.4=h5b20da3_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.9=1_cp39
pytorch=1.8.0=cpu_py39hc766e51_1
readline=8.1=hedafd6a_0
regex=2021.4.4=py39h5161555_0
requests=2.25.1=pyhd3deb0d_0
sacremoses=0.0.43=pyh9f0ad1d_0
setuptools=49.6.0=py39h2804cbe_3
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h27ca646_1
sqlite=3.35.5=hc49ca36_0
tk=8.6.10=hf7e6567_1
tokenizers=0.10.1=py39hda0fb44_0
torchvision=0.2.2=py_3
tornado=6.1=py39h46acfd9_1
tqdm=4.61.0=pyhd8ed1ab_0
transformers=4.6.1=pyhd8ed1ab_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
tzdata=2021a=he74cb21_0
urllib3=1.26.5=pyhd8ed1ab_0
wheel=0.36.2=pyhd3deb0d_0
xz=5.2.5=h642e427_1
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h31e879b_1009
zstd=1.5.0=h861e0a7_0
